requirements.txt
aniso8601==9.0.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.0
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
